Moved classes to models folder in gateways
