+++ HousingSearchApi/V1/Gateways/Models/Identification.cs
@@ -1,4 +1,4 @@
-namespace HousingSearchApi.V1.Gateways
+namespace HousingSearchApi.V1.Gateways.Models
 {
     public class Identification
     {
