+++ HousingSearchApi/V1/Interfaces/ISearchPersonsQueryContainerOrchestrator.cs
@@ -1,5 +1,6 @@
 using HousingSearchApi.V1.Boundary.Requests;
 using HousingSearchApi.V1.Gateways;
+using HousingSearchApi.V1.Gateways.Models;
 using Nest;
 
 namespace HousingSearchApi.V1.Interfaces
