+++ HousingSearchApi/V1/Interfaces/Sorting/LastNameDesc.cs
@@ -1,4 +1,5 @@
 using HousingSearchApi.V1.Gateways;
+using HousingSearchApi.V1.Gateways.Models;
 using Nest;
 
 namespace HousingSearchApi.V1.Interfaces.Sorting
